cpuarch: move MSR and CR0/CR4 definitions to common crate
Move the MSR definitions of EFER and SEV_STATUS as well as the
definitions of CR0 and CR4 bits.  Moving these architcetural definitions to a
common crate means they can be accessed from outside the kernel crate.
