+++ cpuarch/src/lib.rs
@@ -3,5 +3,7 @@
 
 #![no_std]
 
+pub mod sev_status;
 pub mod snp_cpuid;
 pub mod vmsa;
+pub mod x86;
