+++ kernel/src/boot_stage2.rs
@@ -7,15 +7,11 @@
 use bootdefs::kernel_launch::Stage2LaunchInfo;
 use core::arch::global_asm;
 use core::mem::offset_of;
-
-use svsm::{
-    cpu::{
-        efer::EFERFlags,
-        msr::{EFER, SEV_STATUS},
-    },
-    mm::PGTABLE_LVL3_IDX_PTE_SELFMAP,
-    types::PAGE_SIZE,
-};
+use cpuarch::sev_status::MSR_SEV_STATUS;
+use cpuarch::x86::EFERFlags;
+use cpuarch::x86::MSR_EFER;
+use svsm::mm::PGTABLE_LVL3_IDX_PTE_SELFMAP;
+use svsm::types::PAGE_SIZE;
 
 global_asm!(
     r#"
@@ -290,10 +286,10 @@ global_asm!(
     pgtable_end:"#,
     PAGE_SIZE = const PAGE_SIZE,
     PGTABLE_LVL3_IDX_PTE_SELFMAP = const PGTABLE_LVL3_IDX_PTE_SELFMAP,
-    EFER = const EFER,
+    EFER = const MSR_EFER,
     LME = const EFERFlags::LME.bits(),
     NXE = const EFERFlags::NXE.bits(),
-    SEV_STATUS = const SEV_STATUS,
+    SEV_STATUS = const MSR_SEV_STATUS,
     PLATFORM_TYPE_OFF = const offset_of!(Stage2LaunchInfo, platform_type) as u32,
     CPUID_OFF = const offset_of!(Stage2LaunchInfo, cpuid_page) as u32,
     options(att_syntax)
