+++ kernel/src/cpu/control_regs.rs
@@ -9,8 +9,9 @@ use crate::address::{Address, PhysAddr};
 use crate::cpu::features::{cpu_has_smap, cpu_has_smep, cpu_has_umip};
 use crate::cpu::shadow_stack::is_cet_ss_supported;
 use crate::platform::SvsmPlatform;
-use bitflags::bitflags;
 use core::arch::asm;
+use cpuarch::x86::CR0Flags;
+use cpuarch::x86::CR4Flags;
 
 #[inline]
 pub fn cr0_init() {
@@ -103,29 +104,6 @@ pub fn cr4_xsave_enable() {
     }
 }
 
-bitflags! {
-    #[derive(Debug, Clone, Copy)]
-    pub struct CR0Flags: u64 {
-        const PE = 1 << 0;  // Protection Enabled
-        const MP = 1 << 1;  // Monitor Coprocessor
-        const EM = 1 << 2;  // Emulation
-        const TS = 1 << 3;  // Task Switched
-        const ET = 1 << 4;  // Extension Type
-        const NE = 1 << 5;  // Numeric Error
-        const WP = 1 << 16; // Write Protect
-        const AM = 1 << 18; // Alignment Mask
-        const NW = 1 << 29; // Not Writethrough
-        const CD = 1 << 30; // Cache Disable
-        const PG = 1 << 31; // Paging
-    }
-}
-
-impl From<usize> for CR0Flags {
-    fn from(bits: usize) -> Self {
-        CR0Flags::from_bits_truncate(bits as u64)
-    }
-}
-
 #[inline]
 pub fn read_cr0() -> CR0Flags {
     let cr0: u64;
@@ -199,39 +177,6 @@ pub unsafe fn write_cr3(cr3: PhysAddr) {
         asm!("mov %rax, %cr3",
              in("rax") cr3.bits(),
              options(att_syntax));
-    }
-}
-
-bitflags! {
-    #[derive(Debug, Clone, Copy)]
-    pub struct CR4Flags: u64 {
-        const VME       = 1 << 0;  // Virtual-8086 Mode Extensions
-        const PVI       = 1 << 1;  // Protected-Mode Virtual Interrupts
-        const TSD       = 1 << 2;  // Time Stamp Disable
-        const DE        = 1 << 3;  // Debugging Extensions
-        const PSE       = 1 << 4;  // Page Size Extensions
-        const PAE       = 1 << 5;  // Physical-Address Extension
-        const MCE       = 1 << 6;  // Machine Check Enable
-        const PGE       = 1 << 7;  // Page-Global Enable
-        const PCE       = 1 << 8;  // Performance-Monitoring Counter Enable
-        const OSFXSR        = 1 << 9;  // Operating System FXSAVE/FXRSTOR Support
-        const OSXMMEXCPT    = 1 << 10; // Operating System Unmasked Exception Support
-        const UMIP      = 1 << 11; // User Mode Instruction Prevention
-        const LA57      = 1 << 12; // 57-bit linear address
-        const FSGSBASE      = 1 << 16; // Enable RDFSBASE, RDGSBASE, WRFSBASE, and
-                           // WRGSBASE instructions
-        const PCIDE     = 1 << 17; // Process Context Identifier Enable
-        const OSXSAVE       = 1 << 18; // XSAVE and Processor Extended States Enable Bit
-        const SMEP      = 1 << 20; // Supervisor Mode Execution Prevention
-        const SMAP      = 1 << 21; // Supervisor Mode Access Protection
-        const PKE       = 1 << 22; // Protection Key Enable
-        const CET       = 1 << 23; // Control-flow Enforcement Technology
-    }
-}
-
-impl From<usize> for CR4Flags {
-    fn from(bits: usize) -> Self {
-        CR4Flags::from_bits_truncate(bits as u64)
     }
 }
